clear_cmd: bring clear command module implementation
This module lets to set a given time for a command in a group. Then, output messages from that command will be deleted after that time. This is useful to avoid excessive spam, mostly from commands like notes, filters, etc

For example we set notes and 30 seconds in a group. Every time notes command is executed, its output will be deleted after 30 seconds in that group.

Check module help for more details.